refactored: final refresh-token.js
